refactor(NumberValidator): InputValidator를 NumberValidator로 수정
시도 횟수를 검증하는 NumberValidator를 구현하고 ErrorMessage 클래스 삭제
