--- src/main/java/utils/ErrorMessage.java
@@ -1,6 +1,0 @@
-package utils;
-
-public class ErrorMessage {
-    public static final String NOT_NUMBER = "시도 횟수가 숫자가 아닙니다.";
-    public static final String NOT_POSITIVE_NUMBER = "시도 횟수가 양수가 아닙니다.";
-}
