+++ src/test/java/view/InputNumberValidatorTest.java
@@ -1,8 +1,6 @@
-package utils;
+package view;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static utils.ErrorMessage.NOT_NUMBER;
-import static utils.ErrorMessage.NOT_POSITIVE_NUMBER;
 
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
@@ -10,15 +8,13 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class InputValidatorTest {
-
-
+class InputNumberValidatorTest {
 
     static Stream<Arguments> invalidTryNumber() {
         return Stream.of(
-                Arguments.of("배키","시도 횟수가 숫자가 아니다.", NOT_NUMBER),
-                Arguments.of("-1","시도 횟수가 음수이다.", NOT_POSITIVE_NUMBER),
-                Arguments.of(" ","시도 횟수가 빈 문자이다.", NOT_NUMBER)
+                Arguments.of("배키","시도 횟수가 숫자가 아니다.", "시도 횟수가 숫자가 아닙니다."),
+                Arguments.of("-1","시도 횟수가 음수이다.", "시도 횟수가 양수가 아닙니다."),
+                Arguments.of(" ","시도 횟수가 빈 문자이다.", "시도 횟수가 숫자가 아닙니다.")
         );
     }
 
@@ -26,7 +22,7 @@ class InputValidatorTest {
     @DisplayName("시도 횟수 예외 발생")
     @MethodSource("invalidTryNumber")
     public void invalidTryNumberThrowException(String tryNumber, String reason, String errorMessage) {
-        assertThatThrownBy(() -> InputValidator.tryNumberValidator(tryNumber))
+        assertThatThrownBy(() -> InputNumberValidator.validatePositiveNumber(tryNumber))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(errorMessage);
     }
